YtsRoster: don't crash if contact's status arrives before TpContact
Similar to a previously-fixed bug in YtsContact, the roster criticals
and discards a status update if a contact's status is updated before
the TpContact has arrived.

This can be reproduced by killing and restarting MC, then Salut, then
an example service.

Fix it in much the same way as for YtsContact, by remembering the
statuses for later.